# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Sales Summery API
  version: 1.0.0
  description: 소규모 사업장 수입·지출 관리 API
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /financial-records:
    get:
      summary: 기록 기간 검색·필터·페이징
      parameters:
        - { name: startDate, in: query, required: true, schema: { type: string, format: date } }
        - { name: endDate, in: query, required: true, schema: { type: string, format: date } }
        - { name: categoryId, in: query, schema: { type: integer, format: int64 } }
        - { name: transactionType, in: query, schema: { $ref: '#/components/schemas/TransactionType' } }
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: sort, in: query, schema: { type: string, default: 'recordDate,desc' } }
      responses: { '200': { description: 성공 } }
    post:
      summary: 기록 등록
      responses: { '201': { description: 생성 성공 }, '400': { $ref: '#/components/responses/BadRequest' } }
  /financial-records/{recordId}:
    get:
      summary: 기록 상세 조회
      parameters: [ { $ref: '#/components/parameters/RecordId' } ]
      responses: { '200': { description: 성공 }, '404': { description: 기록 없음 } }
    patch:
      summary: 기록 수정 및 항목 변경
      parameters: [ { $ref: '#/components/parameters/RecordId' } ]
      responses: { '200': { description: 성공 } }
    delete:
      summary: 기록 삭제
      parameters: [ { $ref: '#/components/parameters/RecordId' } ]
      responses: { '200': { description: 성공 } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RecordId: { name: recordId, in: path, required: true, schema: { type: integer, format: int64 } }
  schemas:
    TransactionType: { type: string, enum: [INCOME, EXPENSE] }
...
  responses:
    BadRequest: { description: 요청값 오류 }
...